requirements.txt
albumentations==1.3.0
opencv-python==4.7.0.72
pytz==2022.7.1
torchmetrics==0.11.4
torchvision==0.14.1
tqdm==4.65.0
wandb==0.13.11